requirements.txt
pandas==2.2.2
polars==0.20.22
pyarrow==15.0.1
fastparquet==2024.2.0
sqlalchemy==2.0.30
psycopg[binary]==3.1.19
plotly==5.22.0
streamlit==1.35.0
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.2
optuna==3.6.1
shap==0.45.1
mlflow==2.14.1
python-dotenv==1.0.1
pyyaml==6.0.1
kaggle==1.6.14
ruff==0.4.7
pytest==8.2.0
papermill==2.6.0
nbconvert==7.16.4
